Porting translations from 8.0
stock_picking_package_preparation_line: porting translations form 8.0

FIX correctly handle lots on preparation lines

IMP stock_picking_package_preparation: use odoo noupdate

IMP stock_picking_package_preparation: use odoo noupdate
FIX PEP8

REF stock_picking_package_preparation_line: default_get with new API

stock_picking_package_preparation_line: FIX headers, file and model names

stock_picking_package_preparation_line: FIX ensure_one and required field

stock_picking_package_preparation_line: rename product_uom_id

FIX stock_picking_package_preparation_line : Field `product_uom` does not exist

FIX stock_picking_package_preparation_line: AttributeError: 'stock.move' object has no attribute 'product_uom_id'

IMP stock_picking_package_preparation_line comments
