+++ stock_picking_package_preparation_line/i18n/fr.po
@@ -7,15 +7,20 @@ msgid ""
 msgstr ""
 "Project-Id-Version: stock-logistics-workflow (8.0)\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-11-28 01:56+0000\n"
-"PO-Revision-Date: 2015-11-27 13:46+0000\n"
-"Last-Translator: <>\n"
+"POT-Creation-Date: 2016-05-04 02:49+0000\n"
+"PO-Revision-Date: 2016-05-03 14:25+0000\n"
+"Last-Translator: OCA Transbot <transbot@odoo-community.org>\n"
 "Language-Team: French (http://www.transifex.com/oca/OCA-stock-logistics-workflow-8-0/language/fr/)\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
 "Language: fr\n"
 "Plural-Forms: nplurals=2; plural=(n > 1);\n"
+
+#. module: stock_picking_package_preparation_line
+#: model:ir.model,name:stock_picking_package_preparation_line.model_res_company
+msgid "Companies"
+msgstr "Sociétés"
 
 #. module: stock_picking_package_preparation_line
 #: field:stock.picking.package.preparation.line,create_uid:0
@@ -28,6 +33,12 @@ msgid "Created on"
 msgstr "Created on"
 
 #. module: stock_picking_package_preparation_line
+#: field:res.company,default_picking_type_for_package_preparation_id:0
+#: field:stock.config.settings,default_picking_type_for_package_preparation_id:0
+msgid "Default Picking Type used in package preparation"
+msgstr ""
+
+#. module: stock_picking_package_preparation_line
 #: view:stock.picking.package.preparation:stock_picking_package_preparation_line.sppp_line_stock_picking_package_preparation_form
 #: field:stock.picking.package.preparation.line,name:0
 msgid "Description"
@@ -37,6 +48,11 @@ msgstr ""
 #: view:stock.picking.package.preparation:stock_picking_package_preparation_line.sppp_line_stock_picking_package_preparation_form
 #: field:stock.picking.package.preparation,line_ids:0
 msgid "Details"
+msgstr ""
+
+#. module: stock_picking_package_preparation_line
+#: field:stock.picking.package.preparation.line,display_name:0
+msgid "Display Name"
 msgstr ""
 
 #. module: stock_picking_package_preparation_line
@@ -52,11 +68,16 @@ msgid ""
 msgstr ""
 
 #. module: stock_picking_package_preparation_line
-#: code:addons/stock_picking_package_preparation_line/models/stock_picking_package_preparation_line.py:229
-#: code:addons/stock_picking_package_preparation_line/models/stock_picking_package_preparation_line.py:235
+#: code:addons/stock_picking_package_preparation_line/models/stock_picking_package_preparation_line.py:250
+#: code:addons/stock_picking_package_preparation_line/models/stock_picking_package_preparation_line.py:259
 #, python-format
 msgid ""
 "Impossible to create confirmed picking. Please Check products availability!"
+msgstr ""
+
+#. module: stock_picking_package_preparation_line
+#: field:stock.picking.package.preparation.line,__last_update:0
+msgid "Last Modified on"
 msgstr ""
 
 #. module: stock_picking_package_preparation_line
@@ -70,6 +91,11 @@ msgid "Last Updated on"
 msgstr "Last Updated on"
 
 #. module: stock_picking_package_preparation_line
+#: field:stock.picking.package.preparation.line,lot_id:0
+msgid "Lot"
+msgstr "Lot"
+
+#. module: stock_picking_package_preparation_line
 #: view:stock.picking.package.preparation:stock_picking_package_preparation_line.sppp_line_stock_picking_package_preparation_form
 #: field:stock.picking.package.preparation.line,note:0
 msgid "Note"
@@ -77,12 +103,18 @@ msgstr "Note"
 
 #. module: stock_picking_package_preparation_line
 #: model:ir.model,name:stock_picking_package_preparation_line.model_stock_picking_package_preparation
+#: view:stock.config.settings:stock_picking_package_preparation_line.view_stock_picking_package_preparation_line_setting
 msgid "Package Preparation"
 msgstr "Préparation de colis"
 
 #. module: stock_picking_package_preparation_line
 #: model:ir.model,name:stock_picking_package_preparation_line.model_stock_picking_package_preparation_line
 msgid "Package Preparation Line"
+msgstr ""
+
+#. module: stock_picking_package_preparation_line
+#: field:stock.picking.package.preparation,picking_type_id:0
+msgid "Picking Type"
 msgstr ""
 
 #. module: stock_picking_package_preparation_line
