+++ stock_picking_package_preparation_line/views/stock_picking_package_preparation_line.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data noupdate="0">
+<odoo noupdate="0">
 
     <record id="sppp_line_stock_picking_package_preparation_form" model="ir.ui.view">
         <field name="name">sppp.line.stock.picking.package.preparation.form</field>
@@ -15,7 +14,7 @@
                             <field name="name" />
                             <field name="product_id"/>
                             <field name="product_uom_qty" />
-                            <field name="product_uom" groups="product.group_uom"/>
+                            <field name="product_uom_id" groups="product.group_uom"/>
                             <field name="lot_id" groups="stock.group_production_lot"/>
                         </tree>
                         <form>
@@ -24,7 +23,7 @@
                                        attrs="{'readonly': [('move_id', '!=', False)]}"/>
                                 <field name="product_uom_qty"
                                        attrs="{'required': [('product_id', '!=', False)]}"/>
-                                <field name="product_uom" groups="product.group_uom"
+                                <field name="product_uom_id" groups="product.group_uom"
                                        attrs="{'required': [('product_id', '!=', False)], 'readonly': [('move_id', '!=', False)]}"/>
                                 <field name="lot_id"
                                        groups="stock.group_production_lot"
@@ -46,5 +45,4 @@
         </field>
     </record>
 
-</data>
-</openerp>
+</odoo>
